Use common macKey type
